cloud_func/requirements.txt
google-cloud-storage
sklearn
pandas
flask